feat: コンポーネント共通化
