+++ src/screens/blog/index.tsx
@@ -1,26 +1,7 @@
-import { PostList } from '#/components'
-import { StatusBar } from 'expo-status-bar'
-import { ImageBackground, StyleSheet, View } from 'react-native'
+import { GlobalLayout, PostList } from '#/components'
 
 export const BlogScreen = () => (
-  <View style={styles.container}>
-    <ImageBackground source={require('#assets/beer.jpg')} style={styles.background}>
-      <StatusBar style="auto" />
-      <PostList tag="blog" />
-    </ImageBackground>
-  </View>
+  <GlobalLayout>
+    <PostList tag="blog" />
+  </GlobalLayout>
 )
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  background: {
-    width: '100%',
-    flex: 1,
-    justifyContent: 'center',
-  },
-})
